requirements.txt
# 核心依赖
Flask>=2.0.0

# 诊断系统客户端（live模式需要）
requests>=2.25.0
python-dateutil>=2.8.0  # ISO-8601时间解析

# 其他常用依赖（根据项目实际情况添加）
# pandas>=1.3.0
# openpyxl>=3.0.0

flask-cors
serverless-wsgi
